refactor(queries): update imports and remove legacy paths
