--- src/features/app/hooks/use-arena-selection-view.ts
@@ -1,2 +1,0 @@
-export * from '@/queries/advance-settings/arena-selection-query-options';
-export * from '@/queries/advance-settings/use-arena-selection-queries';
