+++ src/features/dashboard/athlete/components/athlete-table/athletes-action-menu.tsx
@@ -32,9 +32,9 @@ import {
   LAST_USED_TOURNAMENT_KEY,
   bulkAddAthleteResult,
 } from '@/features/dashboard/athlete/lib/bulk-add-athletes';
-import { useBulkAddAthletes } from '@/queries/tournament-athletes';
-import { useTournaments } from '@/queries/tournaments';
-import { useUpdateAthleteProfile } from '@/queries/athlete-profiles';
+import { useBulkAddAthletes } from '@/queries/tournament-athlete';
+import { useTournaments } from '@/queries/tournament';
+import { useUpdateAthleteProfile } from '@/queries/athlete-profile';
 
 interface AthletesActionMenuProps {
   options: ColumnOptions;
