+++ src/features/dashboard/tournament/builder/components/brackets-tab/groups-panel/arena-group-order-sheet/arena-group-orders-panel.tsx
@@ -37,7 +37,7 @@ import {
   useRetireArena,
   useSetArenaGroupOrder,
   useTournament,
-} from '@/queries/tournaments';
+} from '@/queries/tournament';
 import { cn } from '@/lib/utils';
 
 interface ArenaGroupOrdersPanelProps {
