+++ src/features/dashboard/tournament/builder/components/dialogs/delete-tournament-dialog.tsx
@@ -9,7 +9,7 @@ import {
   DialogTitle,
 } from '@/components/ui/dialog';
 import { Spinner } from '@/components/ui/spinner';
-import { useDeleteTournament } from '@/queries/tournaments';
+import { useDeleteTournament } from '@/queries/tournament';
 
 interface DeleteTournamentDialogProps {
   open: boolean;
