--- src/lib/queries/matches.ts
@@ -1,1 +1,0 @@
-export * from '@/queries/lib/match/optimistic-updates';
