--- src/lib/queries/tournament-athletes.ts
@@ -1,1 +1,0 @@
-export * from '@/queries/lib/tournament-athlete/list-identity';
