+++ src/orpc/client.ts
@@ -1,7 +1,6 @@
 import { createRouterClient } from '@orpc/server';
 import { createORPCClient } from '@orpc/client';
 import { RPCLink } from '@orpc/client/fetch';
-import { createTanstackQueryUtils } from '@orpc/tanstack-query';
 import { getRequestHeaders } from '@tanstack/react-start/server';
 import { createIsomorphicFn } from '@tanstack/react-start';
 
@@ -25,5 +24,3 @@ const getORPCClient = createIsomorphicFn()
   });
 
 export const client: RouterClient<typeof router> = getORPCClient();
-
-export const orpc = createTanstackQueryUtils(client);
